backend/requirements.txt
django-taggit
django-simple-history
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
python-decouple
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
Pillow==10.4.0
PyPDF2==3.0.1
python-magic==0.4.27
django-storages==1.14.2
boto3==1.34.0
django-allauth==0.57.0
django-otp==1.2.2
django-otp-yubikey==1.0.1
django-extensions==3.2.3
django-debug-toolbar==4.2.0
gunicorn==21.2.0
whitenoise==6.6.0
django-cleanup==8.1.0
django-taggit==4.0.0
django-ckeditor==6.7.0
django-simple-history==3.4.0
django-auditlog==2.3.0
django-ratelimit==4.1.0
django-axes==6.1.1
django-defender==0.9.0
django-silk==5.0.3
django-prometheus==2.3.1
